Moved `randomUniqueString` to util.js. Added `randomString` function to util.js.
